Create helper for array elemennts.
