+++ modules/json_form_widget/json_form_widget.services.yml
@@ -1,10 +1,13 @@
 services:
   json_form.builder:
     class: \Drupal\json_form_widget\JsonFormBuilder
-    arguments: ['@dkan.metastore.schema_retriever','@json_form.string_helper','@json_form.object_helper']
+    arguments: ['@dkan.metastore.schema_retriever','@json_form.string_helper','@json_form.object_helper', '@json_form.array_helper']
   json_form.value_handler:
     class: \Drupal\json_form_widget\JsonFormValueHandler
   json_form.string_helper:
     class: \Drupal\json_form_widget\JsonFormStringHelper
   json_form.object_helper:
     class: \Drupal\json_form_widget\JsonFormObjectHelper
+  json_form.array_helper:
+    class: \Drupal\json_form_widget\JsonFormArrayHelper
+    arguments: ['@json_form.object_helper']
